[CLANG] Fix ARM64 warning cleanup issues
Add missing ARM64 intrinsic declarations and Clang warning gates needed by the build.

Tighten signedness, enum, format, and narrowing conversions reported by Clang across affected user-mode, kernel, WDF, filesystem, and loader code.

Fix the NDIS 6 legacy packet wrapper to use the valid Reserved[0]/Reserved[1] slots instead of indexing past the two-entry Reserved array.

+++ base/applications/atactl/atactl.cpp
@@ -1018,7 +1018,7 @@ ata_adapter_info(
             if(AdapterInfo->AdapterInterfaceType == PCIBus) {
                 slotData.u.AsULONG = AdapterInfo->slotNumber;
                 printf("  PCI Bus/Dev/Func:   %lu/%lu/%lu%s\n",
-                    AdapterInfo->SystemIoBusNumber, slotData.u.bits.DeviceNumber, slotData.u.bits.FunctionNumber,
+                    AdapterInfo->SystemIoBusNumber, (ULONG)slotData.u.bits.DeviceNumber, (ULONG)slotData.u.bits.FunctionNumber,
                     AdapterInfo->AdapterInterfaceType == AdapterInfo->OrigAdapterInterfaceType ? "" : " (ISA-Bridged)");
                 printf("  VendorId/DevId/Rev: %#04x/%#04x/%#02x\n",
                     (USHORT)(AdapterInfo->DevID >> 16),
